create separeted actions modules
